bionemo-recipes/recipes/esm2_minifold_te/requirements.txt
biopython>=1.80
datasets
einops
fair-esm
hydra-core
megatron-fsdp
ml_collections
pytest
scipy
seaborn
torch
torchao!=0.14.0
torchdata
torchmetrics
tqdm
transformer_engine[pytorch]
transformers>=4.44.0
wandb
nvdlfw_inspect @ git+https://github.com/NVIDIA/nvidia-dlfw-inspect
